aws/requirements.txt
# AWS Integration Requirements
boto3==1.34.0
botocore==1.34.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
